webapp/requirements.txt
torch==2.0.1
torchvision==0.15.2
flask==2.3.3
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.74
matplotlib==3.7.2
scikit-learn==1.3.0
pandas==2.0.3
gunicorn==21.2.0